+++ src/main/java/cc/cassian/cauldrons/compat/rrv/brewing/CauldronBrewingClientRecipe.java
@@ -2,6 +2,7 @@ package cc.cassian.cauldrons.compat.rrv.brewing;
 //? if >1.21.10 {
 
 import cc.cassian.cauldrons.compat.rrv.CauldronModRRVPlugin;
+import cc.cassian.cauldrons.compat.rrv.Constants;
 import cc.cassian.cauldrons.core.CauldronContents;
 import cc.cassian.rrv.api.recipe.ReliableClientRecipe;
 import cc.cassian.rrv.api.recipe.ReliableClientRecipeType;
@@ -29,18 +30,18 @@ public class CauldronBrewingClientRecipe implements ReliableClientRecipe {
     @Override
     public void bindSlots(RecipeViewMenu.SlotFillContext slotFillContext) {
         slotFillContext.bindOptionalSlot(0, reagent, RecipeViewMenu.OptionalSlotRenderer.DEFAULT);
-        slotFillContext.bindOptionalSlot(1, CauldronModRRVPlugin.getResultForDisplay(potion).getB(), RecipeViewMenu.OptionalSlotRenderer.DEFAULT);
-        slotFillContext.bindOptionalSlot(2, CauldronModRRVPlugin.getResultForDisplay(result).getB(), RecipeViewMenu.OptionalSlotRenderer.DEFAULT);
+        slotFillContext.bindOptionalSlot(1, Constants.getResultForDisplay(potion).getB(), RecipeViewMenu.OptionalSlotRenderer.DEFAULT);
+        slotFillContext.bindOptionalSlot(2, Constants.getResultForDisplay(result).getB(), RecipeViewMenu.OptionalSlotRenderer.DEFAULT);
     }
 
     @Override
     public List<SlotContent> getIngredients() {
-        return List.of(reagent, CauldronModRRVPlugin.getResultForDisplay(potion).getA());
+        return List.of(reagent, Constants.getResultForDisplay(potion).getA());
     }
 
     @Override
     public List<SlotContent> getResults() {
-        return List.of(CauldronModRRVPlugin.getResultForDisplay(result).getA());
+        return List.of(Constants.getResultForDisplay(result).getA());
     }
 }
 //?}