+++ src/main/java/cc/cassian/cauldrons/compat/rrv/dipping/CauldronDippingViewRecipe.java
@@ -3,6 +3,7 @@ package cc.cassian.cauldrons.compat.rrv.dipping;
 //? if >1.21.10 {
 
 import cc.cassian.cauldrons.compat.rrv.CauldronModRRVPlugin;
+import cc.cassian.cauldrons.compat.rrv.Constants;
 import cc.cassian.cauldrons.core.CauldronContents;
 import cc.cassian.rrv.api.recipe.ReliableClientRecipe;
 import cc.cassian.rrv.api.recipe.ReliableClientRecipeType;
@@ -30,13 +31,13 @@ public class CauldronDippingViewRecipe implements ReliableClientRecipe {
     @Override
     public void bindSlots(RecipeViewMenu.SlotFillContext slotFillContext) {
         slotFillContext.bindOptionalSlot(0, reagent, RecipeViewMenu.OptionalSlotRenderer.DEFAULT);
-        slotFillContext.bindOptionalSlot(1, CauldronModRRVPlugin.getResultForDisplay(potion).getB(), RecipeViewMenu.OptionalSlotRenderer.DEFAULT);
+        slotFillContext.bindOptionalSlot(1, Constants.getResultForDisplay(potion).getB(), RecipeViewMenu.OptionalSlotRenderer.DEFAULT);
         slotFillContext.bindOptionalSlot(2, result, RecipeViewMenu.OptionalSlotRenderer.DEFAULT);
     }
 
     @Override
     public List<SlotContent> getIngredients() {
-        return List.of(reagent, CauldronModRRVPlugin.getResultForDisplay(potion).getA());
+        return List.of(reagent, Constants.getResultForDisplay(potion).getA());
     }
 
     @Override